# Conflicts:
#	taptalk/src/main/java/io/taptalk/TapTalk/View/Adapter/TAPMessageAdapter.java

+++ taptalk/src/main/java/io/taptalk/TapTalk/Helper/TapTalk.java
@@ -32,6 +32,7 @@ import io.taptalk.TapTalk.Listener.TAPDatabaseListener;
 import io.taptalk.TapTalk.Listener.TAPListener;
 import io.taptalk.TapTalk.Manager.TAPCacheManager;
 import io.taptalk.TapTalk.Manager.TAPChatManager;
+import io.taptalk.TapTalk.Manager.TAPConnectionManager;
 import io.taptalk.TapTalk.Manager.TAPCustomBubbleManager;
 import io.taptalk.TapTalk.Manager.TAPDataManager;
 import io.taptalk.TapTalk.Manager.TAPMessageStatusManager;
@@ -166,6 +167,8 @@ public class TapTalk {
                 registerFcmToken();
 
                 TAPDataManager.getInstance().saveActiveUser(response.getUser());
+                TAPApiManager.getInstance().setLogout(false);
+                TAPConnectionManager.getInstance().connect();
                 loginInterface.onLoginSuccess();
             }
 
@@ -198,16 +201,22 @@ public class TapTalk {
         }
     }
 
+
     // TODO: 15/10/18 saat integrasi harus di ilangin
     public static void refreshTokenExpired() {
-        TAPApiManager.getInstance().setLogout(true);
-        TAPRoomListViewModel.setShouldNotLoadFromAPI(false);
-        TAPChatManager.getInstance().disconnectAfterRefreshTokenExpired();
-        TAPDataManager.getInstance().deleteAllPreference();
-        TAPDataManager.getInstance().deleteAllFromDatabase();
-        Intent intent = new Intent(appContext, TAPLoginActivity.class);
-        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-        appContext.startActivity(intent);
+        if (null == tapTalk) {
+            throw new IllegalStateException(appContext.getString(R.string.init_taptalk));
+        } else {
+            TAPApiManager.getInstance().setLogout(true);
+            TAPRoomListViewModel.setShouldNotLoadFromAPI(false);
+            TAPChatManager.getInstance().disconnectAfterRefreshTokenExpired();
+            TAPDataManager.getInstance().deleteAllPreference();
+            TAPDataManager.getInstance().deleteAllFromDatabase();
+
+            for (TAPListener listener : getTapTalkListeners()) {
+                listener.onRefreshTokenExpiredOrInvalid();
+            }
+        }
     }
 
     public static void saveFirebaseToken(String newFirebaseToken) {
